+++ packages/connect-wallet-modal/src/components/WalletsModal/types.ts
@@ -1,12 +1,6 @@
-import React, { ReactNode } from 'react';
+import { ComponentType } from 'react';
 import { ModalProps } from '@reef-knot/ui-react';
 import { WalletAdapterData } from '@reef-knot/types';
-
-export type RequirementsData = {
-  icon?: ReactNode;
-  title?: string;
-  text?: ReactNode;
-};
 
 export type Metrics = {
   events?: {
@@ -16,7 +10,7 @@ export type Metrics = {
 };
 
 export type ButtonComponentsByConnectorId = {
-  [K: string]: React.ComponentType;
+  [K: string]: ComponentType;
 };
 
 export type WalletsModalProps = ModalProps & {
@@ -35,11 +29,5 @@ export type ButtonsCommonProps = {
   onBeforeConnect?: () => void;
   onConnect?: () => void;
   shouldInvertWalletIcon: boolean;
-  setRequirements(
-    this: void,
-    isVisible: boolean,
-    requirementsData: RequirementsData,
-  ): void;
-  setLedgerScreenVisible(this: void, isVisible: boolean): void;
   metrics?: Metrics;
 };