+++ packages/core-react/src/hooks/index.ts
@@ -4,3 +4,5 @@ export * from './useAutoConnectCheck';
 export * from './useEagerConnect';
 export * from './useDisconnect';
 export * from './useConnectorInfo';
+export * from './useConnect';
+export * from './useReefKnotModal';