+++ packages/connect-wallet-modal/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @reef-knot/connect-wallet-modal
 
+## 3.0.0
+
+### Major Changes
+
+- Rework modals interface, add useConnect hook for connection
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/core-react@3.0.0
+  - @reef-knot/web3-react@3.0.0
+
 ## 2.1.0
 
 ### Minor Changes